upgraded from x32 to x64

+++ FlyingCrane3D/GUI.cpp
@@ -204,7 +204,8 @@ void GUI::showSavePanel(bool* showSave) {
 		static int currentSceneItem = 0;
 
 		ImGui::ListBox("Сохраненные\nсцены", &currentSceneItem,
-					   &(sceneItems)[0], sceneItems.size(), 10);
+					   &(sceneItems)[0],
+					   static_cast<int>(sceneItems.size()), 10);
 
 		ImGui::NextColumn();
 
@@ -216,7 +217,8 @@ void GUI::showSavePanel(bool* showSave) {
 		static int currentCategoryItem = 0;
 
 		ImGui::ListBox("Категория", &currentCategoryItem,
-					   &(categoryItems)[0], categoryItems.size(), 8);
+					   &(categoryItems)[0],
+					   static_cast<int>(categoryItems.size()), 8);
 
 		ImGui::NewLine();
 
@@ -278,7 +280,8 @@ void GUI::showLoadPanel(bool* showLoad) {
 		static int currentSceneItem = 0;
 
 		ImGui::ListBox("Сохраненные\nсцены", &currentSceneItem,
-					   &(sceneItems)[0], sceneItems.size(), 10);
+					   &(sceneItems)[0],
+					   static_cast<int>(sceneItems.size()), 10);
 
 		ImGui::NextColumn();
 
@@ -290,7 +293,8 @@ void GUI::showLoadPanel(bool* showLoad) {
 		static int currentCategoryItem = 0;
 
 		ImGui::ListBox("Категория", &currentCategoryItem,
-					   &(categoryItems)[0], categoryItems.size(), 8);
+					   &(categoryItems)[0],
+					   static_cast<int>(categoryItems.size()), 8);
 
 		ImGui::NewLine();
 		ImGui::NewLine();
@@ -504,7 +508,7 @@ void GUI::showToolBar() {
 		ImGui::SetColumnWidth(2, 97.0f);
 		ImGui::SetColumnWidth(3, 97.0f);
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::POINT].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::POINT].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			selectionSwitcher.setSelectionMode(SelectionMode::POINT);
 		}//fi IsItemClicked Button
@@ -514,7 +518,7 @@ void GUI::showToolBar() {
 
 		ImGui::SameLine();
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::EDGE].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::EDGE].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			selectionSwitcher.setSelectionMode(SelectionMode::EDGE);
 		}//fi IsItemClicked Button
@@ -524,7 +528,7 @@ void GUI::showToolBar() {
 
 		ImGui::SameLine();
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::FACE].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::FACE].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			selectionSwitcher.setSelectionMode(SelectionMode::FACE);
 		}//fi IsItemClicked Button
@@ -534,7 +538,7 @@ void GUI::showToolBar() {
 
 		ImGui::SameLine();
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::OBJECT].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::OBJECT].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			selectionSwitcher.setSelectionMode(SelectionMode::OBJECT);
 		}//fi IsItemClicked Button
@@ -544,7 +548,7 @@ void GUI::showToolBar() {
 
 		ImGui::NextColumn();
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::MOVE].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::MOVE].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			if (selection.hasSelectedObjects()) {
 				selectionSwitcher.setActionMode(SelectionActionMode::MOVING);
@@ -556,7 +560,7 @@ void GUI::showToolBar() {
 
 		ImGui::SameLine();
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::ROTATE].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::ROTATE].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			if (selection.hasSelectedObjects()) {
 				selectionSwitcher.setActionMode(SelectionActionMode::ROTATION);
@@ -568,7 +572,7 @@ void GUI::showToolBar() {
 
 		ImGui::SameLine();
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::SCALE].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::SCALE].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			if (selection.hasSelectedObjects()) {
 				selectionSwitcher.setActionMode(SelectionActionMode::SCALING);
@@ -580,7 +584,7 @@ void GUI::showToolBar() {
 
 		ImGui::SameLine(0.0f, spacing_w);
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::REMOVE].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::REMOVE].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			if (selection.hasSelectedObjects()) {
 				selection.deleting();
@@ -592,7 +596,7 @@ void GUI::showToolBar() {
 
 		ImGui::NextColumn();
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::PERSP].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::PERSP].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			scene.getModel().getOrthoProj().switchToPerspective();
 		}//fi IsItemClicked Button
@@ -602,7 +606,7 @@ void GUI::showToolBar() {
 
 		ImGui::SameLine();
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::ORTHO].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::ORTHO].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			if (scene.getCameraSwitcher().getType() == CameraType::TARGET) {
 				scene.getModel().getOrthoProj().switchToOrthographic();
@@ -614,7 +618,7 @@ void GUI::showToolBar() {
 
 		ImGui::NextColumn();
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::CAMERA].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::CAMERA].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			if ((scene.getCameraSwitcher().getType() == CameraType::TARGET &&
 				 scene.getSelectionSwitcher().getActionMode() == SelectionActionMode::NONE) ||
@@ -629,7 +633,7 @@ void GUI::showToolBar() {
 
 		ImGui::SameLine();
 
-		ImGui::ImageButton((ImTextureID)this->icons[GUIIcons::FOCUS].getId(), sizeMoveButton);
+		ImGui::ImageButton((ImTextureID)(__int64)this->icons[GUIIcons::FOCUS].getId(), sizeMoveButton);
 		if (ImGui::IsItemClicked()) {
 			if (scene.getCameraSwitcher().getType() == CameraType::TARGET &&
 				scene.getSelectionSwitcher().getActionMode() == SelectionActionMode::NONE) {
@@ -693,7 +697,8 @@ void GUI::showToolsPanel() {
 
 			ImGui::Separator();
 			ImGui::ListBox("Список\nобъектов", &currentListboxItem, &(listboxItems)[0],
-						   listboxItems.size(), listboxItems.size());
+						   static_cast<int>(listboxItems.size()),
+						   static_cast<int>(listboxItems.size()));
 
 			ImGui::NewLine();
